deploy/requirements.txt
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
functions-framework==3.4.0
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1
db-dtypes==1.1.1
requests==2.31.0
python-dateutil==2.8.2